Query: Refactor shaper processing for relational
- This makes is single pass, end to end processing.
- This is required since split queries need to integrate a processed shaper inside a client method while working on outer shaper.
